+++ OneBlink.SDK/models/FormSubmissionEvent.cs
@@ -70,12 +70,13 @@ namespace OneBlink.SDK.Model
             bool conditionallyExecute = false,
             bool requiresAllConditionallyExecutePredicates = false,
             bool groupFiles = false,
-            List<string> excludedElementIds = default(List<string>),
+            List<Guid> excludedElementIds = default(List<Guid>),
             bool? usePagesAsBreaks = null,
             string pdfFileName = null,
             bool? includeSubmissionIdInPdf = null,
             bool? includePaymentInPdf = null,
-            string label = null
+            string label = null,
+            List<string> excludedCSSClasses = default(List<string>)
             )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -86,11 +87,18 @@ namespace OneBlink.SDK.Model
             fseconfig.actionDefinition = actionDefinition;
             fseconfig.location = location;
             fseconfig.groupFiles = groupFiles;
-            fseconfig.excludedElementIds = excludedElementIds;
+            if (excludedElementIds != default(List<Guid>))
+            {
+                fseconfig.excludedElementIds = excludedElementIds;
+            }
             fseconfig.usePagesAsBreaks = true;
             fseconfig.pdfFileName = pdfFileName;
             fseconfig.includeSubmissionIdInPdf = includeSubmissionIdInPdf;
             fseconfig.includePaymentInPdf = includePaymentInPdf;
+            if (excludedCSSClasses != default(List<string>))
+            {
+                fseconfig.excludedCSSClasses = excludedCSSClasses;
+            }
             FormSubmissionEvent trim = new FormSubmissionEvent();
             trim.type = "TRIM";
             trim.conditionallyExecute = conditionallyExecute;
@@ -123,13 +131,14 @@ namespace OneBlink.SDK.Model
             FormSubmissionEventCivicaRecord civicaCategory,
             List<FormSubmissionEventConfigurationMapping> mapping,
             string pdfFileName = null, Boolean? includeSubmissionIdInPdf = null,
-            List<string> excludedElementIds = default(List<string>),
+            List<Guid> excludedElementIds = default(List<Guid>),
             bool? usePagesAsBreaks = null,
             List<ConditionallyShowPredicate> conditionallyExecutePredicates = default(List<ConditionallyShowPredicate>),
             bool conditionallyExecute = false,
             bool requiresAllConditionallyExecutePredicates = false,
             Boolean? includePaymentInPdf = null,
-            string label = null)
+            string label = null,
+            List<string> excludedCSSClasses = default(List<string>))
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
             fseconfig.environmentId = environmentId;
@@ -139,8 +148,15 @@ namespace OneBlink.SDK.Model
             fseconfig.pdfFileName = pdfFileName;
             fseconfig.includeSubmissionIdInPdf = includeSubmissionIdInPdf;
             fseconfig.includePaymentInPdf = includePaymentInPdf;
-            fseconfig.excludedElementIds = excludedElementIds;
+            if (excludedElementIds != default(List<Guid>))
+            {
+                fseconfig.excludedElementIds = excludedElementIds;
+            }
             fseconfig.usePagesAsBreaks = usePagesAsBreaks;
+            if (excludedCSSClasses != default(List<string>))
+            {
+                fseconfig.excludedCSSClasses = excludedCSSClasses;
+            }
             FormSubmissionEvent civicaCrm = new FormSubmissionEvent();
             civicaCrm.type = "CIVICA_CRM";
             civicaCrm.configuration = fseconfig;
@@ -180,7 +196,7 @@ namespace OneBlink.SDK.Model
             string emailSubjectLine = null,
             string pdfFileName = null,
             bool? includeSubmissionIdInPdf = null,
-            List<string> excludedElementIds = default(List<string>),
+            List<Guid> excludedElementIds = default(List<Guid>),
             bool? usePagesAsBreaks = null,
             FormSubmissionEventEmailTemplate emailTemplate = null,
             List<ConditionallyShowPredicate> conditionallyExecutePredicates = default(List<ConditionallyShowPredicate>),
@@ -190,7 +206,8 @@ namespace OneBlink.SDK.Model
             string label = null,
             List<string> toEmail = default(List<string>),
             List<string> ccEmail = default(List<string>),
-            List<string> bccEmail = default(List<string>)
+            List<string> bccEmail = default(List<string>),
+            List<string> excludedCSSClasses = default(List<string>)
         )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -198,9 +215,16 @@ namespace OneBlink.SDK.Model
             fseconfig.pdfFileName = pdfFileName;
             fseconfig.includeSubmissionIdInPdf = includeSubmissionIdInPdf;
             fseconfig.includePaymentInPdf = includePaymentInPdf;
-            fseconfig.excludedElementIds = excludedElementIds;
+            if (excludedElementIds != default(List<Guid>))
+            {
+                fseconfig.excludedElementIds = excludedElementIds;
+            }
             fseconfig.usePagesAsBreaks = usePagesAsBreaks;
             fseconfig.emailTemplate = emailTemplate;
+            if (excludedCSSClasses != default(List<string>))
+            {
+                fseconfig.excludedCSSClasses = excludedCSSClasses;
+            }
             if (toEmail != default(List<string>))
             {
                 fseconfig.toEmail = toEmail;